Latest improvements to editor

+++ VortexEditor/GUI/VListBox.h
@@ -8,12 +8,12 @@ public:
   VListBox();
   VListBox(HINSTANCE hinstance, VWindow &parent, const std::string &title, 
     COLORREF backcol, uint32_t width, uint32_t height, uint32_t x, uint32_t y,
-    uint32_t menuID, VWindowCallback callback);
+    uintptr_t menuID, VWindowCallback callback);
   virtual ~VListBox();
 
   virtual void init(HINSTANCE hinstance, VWindow &parent, const std::string &title, 
     COLORREF backcol, uint32_t width, uint32_t height, uint32_t x, uint32_t y,
-    uint32_t menuID, VWindowCallback callback);
+    uintptr_t menuID, VWindowCallback callback);
   virtual void cleanup();
 
   // windows message handlers
